Make all ChatConfig fields uint32:     Because they do not need to support such large numbers anyways.

+++ server/cmd/api/chatconfig.go
@@ -2,7 +2,7 @@ package main
 
 type ChatConfig struct {
 	ChatRoomTTL                 uint32 `json:"chatRoomTTL"`
-	MaxNumberOfParticipants     uint64 `json:"maxNumberOfParticipants"`
-	MaxHistoryDurationInSeconds uint64 `json:"maxHistoryDurationInSeconds"`
-	MaxMessagesLength           uint64 `json:"maxMessagesLength"`
+	MaxNumberOfParticipants     uint32 `json:"maxNumberOfParticipants"`
+	MaxHistoryDurationInSeconds uint32 `json:"maxHistoryDurationInSeconds"`
+	MaxMessagesLength           uint32 `json:"maxMessagesLength"`
 }
